+++ public/index.html
@@ -1,48 +1,38 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link
-      href="https://fonts.googleapis.com/css?family=Open+Sans:400,700,800&amp;subset=cyrillic-ext"
-      rel="stylesheet"
-    />
-    <link
-      rel="stylesheet"
-      type="text/css"
-      media="screen"
-      href="./css/normalize.css"
-    />
-    <link
-      rel="stylesheet"
-      type="text/css"
-      media="screen"
-      href="./css/main.css"
-    />
-    <title>Cinemaddict</title>
-  </head>
-  <body>
-    <header class="header">
-      <h1 class="header__logo logo">Cinemaddict</h1>
 
-      <!-- Звание пользователя -->
-    </header>
+<head>
+  <meta charset="utf-8" />
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <link href="https://fonts.googleapis.com/css?family=Open+Sans:400,700,800&amp;subset=cyrillic-ext" rel="stylesheet" />
+  <link rel="stylesheet" type="text/css" media="screen" href="./css/normalize.css" />
+  <link rel="stylesheet" type="text/css" media="screen" href="./css/main.css" />
+  <title>Cinemaddict</title>
+</head>
 
-    <main class="main">
-      <!-- Меню -->
+<body>
+  <header class="header">
+    <h1 class="header__logo logo">Cinemaddict</h1>
 
-      <!-- Сортировка -->
+    <!-- Звание пользователя -->
+  </header>
 
-      <!-- Контент -->
-    </main>
+  <main class="main">
+    <!-- Меню -->
 
-    <footer class="footer">
-      <section class="footer__logo logo logo--smaller">Cinemaddict</section>
-      <section class="footer__statistics">
-        <!-- Количество фильмов -->
-      </section>
-    </footer>
-    <script src="../src/main.js"></script>
-  </body>
+    <!-- Сортировка -->
+
+    <!-- Контент -->
+  </main>
+
+  <footer class="footer">
+    <section class="footer__logo logo logo--smaller">Cinemaddict</section>
+    <section class="footer__statistics">
+      <!-- Количество фильмов -->
+    </section>
+  </footer>
+  <script src="./bundle.js"></script>
+</body>
+
 </html>
