+++ src/main.js
@@ -1,114 +1,31 @@
-const render = (container, el, position = "beforeend") => {
-  container.insertAdjacentHTML(position, el);
-};
-const profile = () => {
-  return `<section class="header__profile profile">
-  <p class="profile__rating">Movie Buff</p>
-  <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-  </section>`;
-};
+import {render} from './util/util';
+import {card} from './view/card';
+import {profile} from './view/profile';
+import {menu} from './view/menu';
+import {layout} from './view/main-layout';
+import {sort} from './view/sort';
+import {button} from './view/button';
+import {filmsLayout} from './view/extra-layout';
 
-const menu = () => {
-  return `
-    <nav class="main-navigation">
-      <div class="main-navigation__items">
-        <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
-        <a href="#watchlist" class="main-navigation__item">Watchlist <span class="main-navigation__item-count">13</span></a>
-        <a href="#history" class="main-navigation__item">History <span class="main-navigation__item-count">4</span></a>
-        <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">8</span></a>
-      </div>
-      <a href="#stats" class="main-navigation__additional">Stats</a>
-    </nav>
-  `;
-};
-
-const sort = () => {
-  return `
-    <ul class="sort">
-      <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
-      <li><a href="#" class="sort__button">Sort by date</a></li>
-      <li><a href="#" class="sort__button">Sort by rating</a></li>
-    </ul>
-  `;
-};
-
-const card = () => {
-  return `
-    <article class="film-card">
-      <h3 class="film-card__title">The Dance of Life</h3>
-      <p class="film-card__rating">8.3</p>
-      <p class="film-card__info">
-        <span class="film-card__year">1929</span>
-        <span class="film-card__duration">1h 55m</span>
-        <span class="film-card__genre">Musical</span>
-      </p>
-      <img src="./images/posters/the-dance-of-life.jpg" alt="" class="film-card__poster">
-      <p class="film-card__description">Burlesque comic Ralph "Skid" Johnson (Skelly), and specialty dancer Bonny Lee King (Carroll), end up together on a cold, rainy night at a tr…</p>
-      <a class="film-card__comments">5 comments</a>
-      <form class="film-card__controls">
-        <button class="film-card__controls-item button film-card__controls-item--add-to-watchlist">Add to watchlist</button>
-        <button class="film-card__controls-item button film-card__controls-item--mark-as-watched">Mark as watched</button>
-        <button class="film-card__controls-item button film-card__controls-item--favorite">Mark as favorite</button>
-      </form>
-    </article>
-  `;
-};
-const layout = () => {
-  return `
-  <section class="films">
-  </section>
-  `;
-};
-const filmsLayout = (classBlock = "", classTitle = "", title) => {
-  return `
-    <section class="films-list${classBlock}">
-      <h2 class="films-list__title ${classTitle}">${title}</h2>
-      <div class="films-list__container">
-      </div>
-    </section>
-  `;
-};
-
-const button = () => {
-  return `
-  <button class="films-list__show-more">Show more</button>
-  `;
-};
-
-const header = document.querySelector(".header");
-const main = document.querySelector(".main");
+const header = document.querySelector(`.header`);
+const main = document.querySelector(`.main`);
 
 render(header, profile());
 render(main, menu());
 render(main, sort());
 render(main, layout());
-const mainContent = document.querySelector(".films");
-render(
-  mainContent,
-  filmsLayout(" film__list_main", "visually-hidden", "All movies. Upcoming")
-);
-render(
-  mainContent,
-  filmsLayout("--extra film__list_top-rated", "", "Top rated")
-);
-render(
-  mainContent,
-  filmsLayout("--extra film__list_most-commented", "", "Most commented")
-);
-const mainFilmContainer = document.querySelector(
-  ".film__list_main .films-list__container"
-);
-const topRatedFilmContainer = document.querySelector(
-  ".film__list_top-rated .films-list__container"
-);
-const mostCommentedFilmContainer = document.querySelector(
-  ".film__list_most-commented .films-list__container"
-);
+const mainContent = document.querySelector(`.films`);
+render(mainContent, filmsLayout(` film__list_main`, `visually-hidden`, `All movies. Upcoming`));
+render(mainContent, filmsLayout(`--extra film__list_top-rated`, ``, `Top rated`));
+render(mainContent, filmsLayout(`--extra film__list_most-commented`, ``, `Most commented`));
+const mainFilmContainer = document.querySelector(`.film__list_main .films-list__container`);
+const topRatedFilmContainer = document.querySelector(`.film__list_top-rated .films-list__container`);
+const mostCommentedFilmContainer = document.querySelector(`.film__list_most-commented .films-list__container`);
 
 for (let i = 0; i < 5; i++) {
   render(mainFilmContainer, card());
 }
-render(mainFilmContainer, button(), "afterend");
+render(mainFilmContainer, button(), `afterend`);
 for (let i = 0; i < 2; i++) {
   render(topRatedFilmContainer, card());
 }
