+++ src/util/util.js
@@ -1,0 +1,4 @@
+const render = (container, el, position = `beforeend`) => {
+  container.insertAdjacentHTML(position, el);
+};
+export {render};
