+++ src/view/button.js
@@ -1,0 +1,9 @@
+const button = () => {
+  return `
+  <button class="films-list__show-more">Show more</button>
+  `;
+};
+
+export {
+  button
+};
