+++ src/view/main-layout.js
@@ -1,0 +1,10 @@
+const layout = () => {
+  return `
+  <section class="films">
+  </section>
+  `;
+};
+
+export {
+  layout
+};
